Rework organize imports command to tolerate errors in ST code
