+++ plugins/org.eclipse.fordiac.ide.gef/META-INF/MANIFEST.MF
@@ -31,6 +31,7 @@ Require-Bundle: org.eclipse.gef,
  org.eclipse.fordiac.ide.model.commands,
  org.eclipse.fordiac.ide.util,
  org.eclipse.fordiac.ide.systemmanagement,
+ org.eclipse.fordiac.ide.typemanagement,
  org.eclipse.fordiac.ide.ui,
  org.eclipse.ui.views.properties.tabbed,
  org.eclipse.emf.edit,
