+++ plugins/org.eclipse.fordiac.ide.typemanagement/META-INF/MANIFEST.MF
@@ -29,6 +29,7 @@ Bundle-ManifestVersion: 2
 Bundle-SymbolicName: org.eclipse.fordiac.ide.typemanagement;singleton:=true
 Bundle-RequiredExecutionEnvironment: JavaSE-21
 Export-Package: org.eclipse.fordiac.ide.typemanagement,
+ org.eclipse.fordiac.ide.typemanagement.commands,
  org.eclipse.fordiac.ide.typemanagement.handlers,
  org.eclipse.fordiac.ide.typemanagement.navigator,
  org.eclipse.fordiac.ide.typemanagement.preferences,
